# Optimize `mergeProps` and `splitProps`