--- a/.changeset/tender-grapes-exercise.md
+++ b/.changeset/tender-grapes-exercise.md
@@ -1,0 +1,5 @@
+---
+"solid-js": patch
+---
+
+Improve performance in `splitProps` and `mergeProps`