@ryansolid 2023-05-01
Interesting. I didn't realize that, so getting property descriptor n times in a loop is faster than that getting all them all at once in the plural form. That's good to know.