@juanrgm 2023-05-11
With the latest changes the tests are successful (I still need to add more tests for `splitProps`).

Not generating breaking changes means a loss of performance, even so the improvement is noticeable (x2).

In the next merge request you can discuss which improvements to unlock (commented out in the code with `// [breaking && performance]`), that's when performance can reach x8.

<details>
<summary>report</summary>

```

 RUN  v0.29.3

 ✓ test/component.bench.ts  (54 tests) 32116ms


 BENCH  Summary

  mergeProps - test/component.bench.ts > mergeProps-static(5, 1)
    7.34x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-static(5, 1, 2)
    8.00x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-static(0, 15)
    8.78x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-static(0, 3, 2)
    6.13x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-static(0, 100)
    6.98x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-static(0, 100, 3, 2)
    6.48x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-static(25, 100)
    5.92x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-static(50, 100)
    4.73x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-static(100, 25)
    6.28x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-signal(5, 1)
    1.59x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-signal(5, 1, 2)
    1.52x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-signal(0, 15)
    1.46x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-signal(0, 3, 2)
    1.53x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-signal(0, 100)
    1.60x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-signal(0, 100, 3, 2)
    1.55x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-signal(25, 100)
    1.51x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-signal(50, 100)
    1.49x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-signal(100, 25)
    1.51x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-mixed(5, 1)
    1.88x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-mixed(5, 1, 2)
    2.40x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-mixed(0, 15)
    2.25x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-mixed(0, 3, 2)
    1.75x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-mixed(0, 100)
    2.35x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-mixed(0, 100, 3, 2)
    2.22x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-mixed(25, 100)
    2.20x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-mixed(50, 100)
    2.38x faster than oldMergeProps

  mergeProps - test/component.bench.ts > mergeProps-mixed(100, 25)
    2.27x faster than oldMergeProps
```
</details>

Before merging this branch, I'll remove the `/component.old.ts` file.