@ryansolid 2023-05-26
This looks good.. can you expand on the `breaking/performance` comments. I'm gathering you are saying it would be breaking but allow better performance. And breaking in what sense? Almost all the changes also reduce code so I'm definitely interested.