@juanrgm 2023-05-27
Uncommenting the `[breaking && performance]` lines will break the following tests in `mergeProps`:
- keeps references
- with getter keeps references
- always returns a new reference
- always returns getters
- always returns unwritables

But it will alsobincrease performance.

The conservative option it would be to not break the tests in `1.7.4` and wait for `1.8.0`. 

That tests are touching the limits and breaking them should not cause problems in the ecosystem because are undocumented features, but I don't know. What do you think?